# Add hybrid `p256_montjadd` and `p256_montjdouble` for Arm, slow multipliers