This patch adds a hybrid (scalar + SIMD) version of `p256_montjadd` and `p256_montjdouble` for Arm with slow multipliers.
This uses the hybrid field operations (`bignum_montmul_p256_neon`, `bignum_montsqr_p256_neon`).

Original:
```
p256_montjadd                   :   513.2 ns each (var  0.1%, corr  0.04) =    1948395 ops/sec
p256_montjdouble                :   276.3 ns each (var  0.1%, corr  0.01) =    3619774 ops/sec
```
New (this patch): both are 11% faster
```
p256_montjadd                   :   461.4 ns each (var  0.1%, corr  0.13) =    2167340 ops/sec
p256_montjdouble                :   248.4 ns each (var  0.3%, corr  0.06) =    4025645 ops/sec
```
To prove this, a clean call-based version is written and placed at `arm/p256/unopt` . The assembly is compiled only when `make proofs` is run, and not included in the library.
The optimized version is placed in `arm/p256` and it is the result of applying function inlining + eliminating redundant load/store instructions + folding (add addr, const) + load/store.

Orthogonal changes are:
- `common/actions_merger.ml` adds OCaml functions that can be used in proofs to combine two lists of actions that are created from, say, diff(fileA,fileB) and diff(fileA,fileC). The combined action corrsponds to diff(fileA,fileC).
- Added `extra_early_rewrite_rules` which has a set of rewrite rules to apply at the early stage of `ASSEMBLER_SIMPLIFY_TAC`. This is helpful when one wants to break memory read of 128 bit into `word_join` of two 64 bit reads.
- `tools/diff.py` now does not use the Python `difftools` package because the package sometimes generate suboptimal diffs. Instead, it invokes the Linux `diff` commandline tool and uses its standard output. Also, its name is changed to `gen-actions.py`.
- Added a cache to `COMPONENT_CANON_CONV` to avoid repetitive computation. This brings ~2% speedup to proofs
- Many other improvements and equivalence checking tactics