@aqjune-aws 2024-07-27
Hoping that CI check is happy this time...

The equiv. check between unoptimized p256_montjadd and optimized p256_montjadd takes slightly less than 2 hours. This is longer than 1 hr + alpha that I introduced before.
The reason is like this. For faster eq. checking, I initially added a custom orthogonal_compnents cache to `ORTHOGONAL_COMPONENTS_CONV` in order to avoid reproving `orthogonal_components` between one memory access and one register. This was used by p256_montjadd's equiv checker to speed up symbolic simulation further. However, I removed this cache later because this solution looked a bit hacky. It seems that more fundamental problem is that a set of MAYCHANGE of p256_montjadd grows big during simulation because the number of words accessed in memory are accessed. Once it grows big, a single instruction can take up to 1 second which is bad. :( I think a better solution is to merge memory locations that are adjacent in MAYCHANGE so that the number of entries is kept small.