@aqjune-aws 2024-07-27
CI check: proof for p256_montjadd passes, the remaining failure is from scalar multiplication proof (specifically `P256_SCALARMUL_SUBROUTINE_CORRECT`). Fixing this doesn't seem hard, but since #135 is also adding base point scalar multiplication it seems to be a good idea to wait until #135 is merged.