@aqjune-aws 2024-08-08
Updated p256_scalarmul to use the optimized point ops & field ops.
TODO is to update p256_scalarmulbase to use the optimized field ops.